+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ### Backwards incompatible changes
   - Layout must now access the child contents with `@inner_content` instead of invoking the LiveView directly
+  - All LiveView callbacks which previously supported returning a `:stop` tuple are no longer supported. LiveViews are stopped when issuing a `redirect` or `push_redirect`
 
 ### Enhancements
+  - Allow `rediret` and `push_rediret` from mount
   - Add `transport_pid/1` to return the websocket transport pid when the socket is connected
 
 ### Bug Fixes