+++ CHANGELOG.md
@@ -5,11 +5,13 @@
   - Returning `:stop` tuples from LiveView `mount` or `handle_[params|call|cast|info]` is no longer supported. LiveViews are stopped when issuing a `redirect` or `push_redirect`
 
 ### Enhancements
+  - Allow `redirect` and `push_redirect` from mount
   - Use acknowledgement tracking to avoid patching inputs until the server has processed the form event
   - Add css loading states to all phx bound elements with event specfic css classes
   - Dispatch `phx:page-loading-start` and `phx:page-loading-stop` on window for live navigation, initial page loads, and form submits, for user controlled page loading integration
   - Allow any phx bound element to specify `phx-page-loading` to dispatch loading events above when the event is pushed
-  - Allow `redirect` and `push_redirect` from mount
+  - Add client side latency simulator with new `enableLatencySim(milliseconds)` and `disableLatencySim()`
+  - Add `enableDebug()` and `disableDebug()` to `LiveSocket` for ondemand browser debugging from the web console
   - Add `transport_pid/1` to return the websocket transport pid when the socket is connected
 
 ### Bug Fixes